Initial commit
The Respoke Modules for Asterisk (RMA) is a collection of libraries
that provide the necessary tools in order to allow communication
between Asterisk and the Respoke service.

The RMA is made up of several installable modules which can be found
under a directory structure similar to Asterisk. For instance, the
channel driver is found under the 'channels' directory and the resource
modules (res_respoke) are contained in the 'res' directory.

The channel driver/protocol stack requires Asterisk (>= version 13.0.0)
as well as libcurl.

A 'respoke.conf' configuration file also needs to be constructed and saved to the Asterisk installation configuration directory (typically /etc/asterisk/). The respoke configuration file follows the same rules and similar patterns to that of a typical Asterisk configuration file. See the 'respoke.conf.sample' file for more information.