res/res_respoke_session: Don't hangup a channel that is already gone
If things become out of sync with the Respoke service, there is a chance that
we will receive an 'error' message after a channel has already been hung up
and disassociated with the session object. If that is the case, don't try
to hang up a NULL channel. Hilarity in the form of FRACKs and crashing will
ensue.