Add support for retrieving metadata from channel
This patch adds a new channel function, `RESPOKE_METADATA([<key>])`, to
allow data to be retrieved from a call coming in from a Respoke
endpoint. If no metadata is present on the call, the function will print
a debug message indicating as such. If no key is provided, the function
will return the value of the metadata, which will be a JSON string if
the metadata is an object or an array. If a key is provided and the
metadata is an object that contains that key, the value of the key will
be provided in the same manner.

Related to MER-4452.
Paired with David Lee.