Prevent make_version overwriting valid .version
Currently when a release is created with `make dist`, it generates
a .version file that is meant to be used by the build process on a
machine that is compiling chan_respoke from an extracted tar
instead of a git checkout. The build_tools/make_version script
takes into consideration an existing .version file, but will
incorrectly overwrite it with a value of 'unknown' when the build
is being run from an extracted tar.

This patch modifies the make_version script to prevent overwriting
an existing .version file with a value of 'unknown'.

Closes #21.