Document release process on non-Asterisk box
When cutting a new release, part of the process is generating
a tar file that can be attached to the github release that has
the appropriate .version file referencing the git tag. The normal
way of generating this tar file is with the `make dist` command.
However, `make dist` also checks for Asterisk to be installed,
since the Makefile is primarily a build tool.

You don't need Asterisk to be installed to build the release tar
file, so this patch adds documentation on how to build it without
using the make target.

Closes #22.