res/res_socket_io: Clarify some ERROR messages, add a debug message
If a bad URL is passed up to Respoke containing an invalid app secret,
the response back is going to consist of a text error message. This
causes the ERROR messages to be a bit confusing, as they will say
something along the lines of "Failed to find session ID in heartbeat
error" - which doesn't tell you that 'heartbeat error' is the actual
response.

This patch clarifies that by putting some single quotes around the
returned response. It also adds some debug that shows the URL that we
send to Respoke, which is necessary to figure out why we might be
getting ERROR messages.