res/res_socket_io_websocket: Don't use protobyname
The function 'protobyname' is not thread safe. In later versions of gcc,
the function has been macro'd to an invalid function call in order to
prevent compilation/usage. This patch replaces the call to protobyname
with direct usage of the constant 'IPPROTO_TCP', which is the protocol
level protobyname was helping us get to.